+++ extension/slim/ModelSlim.java
@@ -35,41 +35,41 @@ public class ModelSlim extends ModelBase implements IModelWithHands
         this.textureHeight = 64;
     
         this.head = new ModelRenderer(this, 0, 0).setTextureSize(this.textureWidth, this.textureHeight);
-        this.head.addBox(-4F, -8F, -4F, 8, 8, 8);
-        this.head.setRotationPoint(0F, -0.2F, 0F);
+        this.head.addBox(-4f, -8f, -4f, 8, 8, 8);
+        this.head.setRotationPoint(0f, -0.2f, 0f);
         this.headWear = new ModelRenderer(this, 0, 16).setTextureSize(this.textureWidth, this.textureHeight);
-        this.headWear.addBox(-4F, -8F, -4F, 8, 8, 8, 0.5f);
-        this.headWear.setRotationPoint(0F, 0F, 0F);
+        this.headWear.addBox(-4f, -8f, -4f, 8, 8, 8, 0.5f);
+        this.headWear.setRotationPoint(0f, 0f, 0f);
         this.body = new ModelRenderer(this, 0, 32).setTextureSize(this.textureWidth, this.textureHeight);
-        this.body.addBox(-4F, 0F, -2F, 8, 16, 4);
-        this.body.setRotationPoint(0F, -8F, 0F);
+        this.body.addBox(-4f, 0f, -2f, 8, 16, 4);
+        this.body.setRotationPoint(0f, -8f, 0f);
         this.neck = new ModelRenderer(this, 0, 52).setTextureSize(this.textureWidth, this.textureHeight);
-        this.neck.addBox(-1.5F, -3F, -1.5F, 3, 3, 3);
-        this.neck.setRotationPoint(0F, 0F, 0F);
+        this.neck.addBox(-1.5f, -3f, -1.5f, 3, 3, 3);
+        this.neck.setRotationPoint(0f, 0f, 0f);
         this.rightUpperArm = new ModelRenderer(this, 32, 0).setTextureSize(this.textureWidth, this.textureHeight);
-        this.rightUpperArm.addBox(-2F, -1.8F, -2F, 4, 8, 4);
-        this.rightUpperArm.setRotationPoint(-6F, -6F, 0F);
+        this.rightUpperArm.addBox(-2f, -1.8f, -2f, 4, 8, 4);
+        this.rightUpperArm.setRotationPoint(-6f, -6f, 0f);
         this.rightLowerArm = new ModelRenderer(this, 32, 12).setTextureSize(this.textureWidth, this.textureHeight);
-        this.rightLowerArm.addBox(-2F, 0F, -2F, 4, 8, 4);
-        this.rightLowerArm.setRotationPoint(0F, 6.2F, 0F);
+        this.rightLowerArm.addBox(-2f, 0f, -2f, 4, 8, 4);
+        this.rightLowerArm.setRotationPoint(0f, 6.2f, 0f);
         this.leftUpperArm = new ModelRenderer(this, 48, 0).setTextureSize(this.textureWidth, this.textureHeight);
-        this.leftUpperArm.addBox(-2F, -1.8F, -2F, 4, 8, 4);
-        this.leftUpperArm.setRotationPoint(6F, -6F, 0F);
+        this.leftUpperArm.addBox(-2f, -1.8f, -2f, 4, 8, 4);
+        this.leftUpperArm.setRotationPoint(6f, -6f, 0f);
         this.leftLowerArm = new ModelRenderer(this, 48, 12).setTextureSize(this.textureWidth, this.textureHeight);
-        this.leftLowerArm.addBox(-2F, 0F, -2F, 4, 8, 4);
-        this.leftLowerArm.setRotationPoint(0F, 6.2F, 0F);
+        this.leftLowerArm.addBox(-2f, 0f, -2f, 4, 8, 4);
+        this.leftLowerArm.setRotationPoint(0f, 6.2f, 0f);
         this.rightUpperLeg = new ModelRenderer(this, 32, 24).setTextureSize(this.textureWidth, this.textureHeight);
-        this.rightUpperLeg.addBox(-2F, 0F, -2F, 4, 8, 4);
-        this.rightUpperLeg.setRotationPoint(-2F, 8F, 0F);
+        this.rightUpperLeg.addBox(-2f, 0f, -2f, 4, 8, 4);
+        this.rightUpperLeg.setRotationPoint(-2f, 8f, 0f);
         this.rightLowerLeg = new ModelRenderer(this, 32, 36).setTextureSize(this.textureWidth, this.textureHeight);
-        this.rightLowerLeg.addBox(-2F, 0F, -2F, 4, 8, 4);
-        this.rightLowerLeg.setRotationPoint(0F, 8F, 0F);
+        this.rightLowerLeg.addBox(-2f, 0f, -2f, 4, 8, 4);
+        this.rightLowerLeg.setRotationPoint(0f, 8f, 0f);
         this.leftUpperLeg = new ModelRenderer(this, 48, 24).setTextureSize(this.textureWidth, this.textureHeight);
-        this.leftUpperLeg.addBox(-2F, 0F, -2F, 4, 8, 4);
-        this.leftUpperLeg.setRotationPoint(2F, 8F, 0F);
+        this.leftUpperLeg.addBox(-2f, 0f, -2f, 4, 8, 4);
+        this.leftUpperLeg.setRotationPoint(2f, 8f, 0f);
         this.leftLowerLeg = new ModelRenderer(this, 48, 36).setTextureSize(this.textureWidth, this.textureHeight);
-        this.leftLowerLeg.addBox(-2F, 0F, -2F, 4, 8, 4);
-        this.leftLowerLeg.setRotationPoint(0F, 8F, 0F);
+        this.leftLowerLeg.addBox(-2f, 0f, -2f, 4, 8, 4);
+        this.leftLowerLeg.setRotationPoint(0f, 8f, 0f);
 
         this.body.addChild(this.head);
         this.body.addChild(this.neck);
@@ -95,10 +95,10 @@ public class ModelSlim extends ModelBase implements IModelWithHands
     public void setRotationAngles(float f, float f1, float f2, float f3, float f4, float f5, Entity entity)
     {
         EntityNpc npc = (EntityNpc)entity;
-        this.head.rotateAngleY = f3 / 57.29578F;
-        this.head.rotateAngleX = f4 / 57.29578F;
+        this.head.rotateAngleY = f3 / 57.29578f;
+        this.head.rotateAngleX = f4 / 57.29578f;
         float naturalArmAngle = 0.01f;
-        float rightArmAngle = MathHelper.cos(f * 0.6662F + 3.141593F) * 2.0F * f1 * 0.5F;
+        float rightArmAngle = MathHelper.cos(f * 0.6662f + 3.141593f) * 2.0f * f1 * 0.5f;
         if ( rightArmAngle < 0 )
         {
             this.rightUpperArm.rotateAngleX = rightArmAngle + naturalArmAngle;
@@ -109,7 +109,7 @@ public class ModelSlim extends ModelBase implements IModelWithHands
             this.rightUpperArm.rotateAngleX = rightArmAngle + naturalArmAngle;
             this.rightLowerArm.rotateAngleX = - naturalArmAngle * 3;
         }
-        float leftArmAngle = MathHelper.cos(f * 0.6662F) * 2.0F * f1 * 0.5F;
+        float leftArmAngle = MathHelper.cos(f * 0.6662f) * 2.0f * f1 * 0.5f;
         if ( leftArmAngle < 0 )
         {
             this.leftUpperArm.rotateAngleX = leftArmAngle + naturalArmAngle;
@@ -126,54 +126,54 @@ public class ModelSlim extends ModelBase implements IModelWithHands
         this.leftUpperArm.rotateAngleY = 0f;
         this.leftLowerArm.rotateAngleY = 0f;
         this.leftUpperArm.rotateAngleZ = 0f;
-        this.rightUpperLeg.rotateAngleX = MathHelper.cos(f * 0.6662F) * 1.0F * f1;
-        this.leftUpperLeg.rotateAngleX = MathHelper.cos(f * 0.6662F + 3.141593F) * 1.0F * f1;
+        this.rightUpperLeg.rotateAngleX = MathHelper.cos(f * 0.6662f) * 1.0f * f1;
+        this.leftUpperLeg.rotateAngleX = MathHelper.cos(f * 0.6662f + 3.141593f) * 1.0f * f1;
         this.rightUpperLeg.rotateAngleY = 0f;
         this.leftUpperLeg.rotateAngleY = 0f;
 
-        if ( onGround > -9990F )
+        if ( onGround > -9990f )
         {
             if ( (npc.getForm() != null) && (npc.getDominantHand().equals(Utility.leftHanded)) )
             {
                 float f6 = onGround;
-                this.body.rotateAngleY = -MathHelper.sin(MathHelper.sqrt_float(f6) * 3.141593F * 2.0F) * 0.2F;
-                this.rightUpperArm.rotationPointZ = MathHelper.sin(this.body.rotateAngleY) * 2.5F;
-                this.leftUpperArm.rotationPointZ = -MathHelper.sin(this.body.rotateAngleY) * 2.5F;
+                this.body.rotateAngleY = -MathHelper.sin(MathHelper.sqrt_float(f6) * 3.141593f * 2.0f) * 0.2f;
+                this.rightUpperArm.rotationPointZ = MathHelper.sin(this.body.rotateAngleY) * 2.5f;
+                this.leftUpperArm.rotationPointZ = -MathHelper.sin(this.body.rotateAngleY) * 2.5f;
                 this.rightUpperArm.rotateAngleY += this.body.rotateAngleY;
                 this.leftUpperArm.rotateAngleY += this.body.rotateAngleY;
                 this.rightUpperArm.rotateAngleX -= this.body.rotateAngleY;
-                f6 = 1.0F - onGround;
+                f6 = 1.0f - onGround;
                 f6 *= f6;
                 f6 *= f6;
-                f6 = 1.0F - f6;
-                float f8 = MathHelper.sin(f6 * 3.141593F);
-                float f10 = MathHelper.sin(onGround * 3.141593F) * -(this.head.rotateAngleX - 0.7F) * 0.75F;
-                this.leftUpperArm.rotateAngleX -= (double)f8 * 1.2D + (double)f10;
-                this.leftUpperArm.rotateAngleY += this.body.rotateAngleY * 2.0F;
-                this.leftUpperArm.rotateAngleZ = -MathHelper.sin(onGround * 3.141593F) * -0.4F;
+                f6 = 1.0f - f6;
+                float f8 = MathHelper.sin(f6 * 3.141593f);
+                float f10 = MathHelper.sin(onGround * 3.141593f) * -(this.head.rotateAngleX - 0.7f) * 0.75f;
+                this.leftUpperArm.rotateAngleX -= f8 * 1.2d + f10;
+                this.leftUpperArm.rotateAngleY += this.body.rotateAngleY * 2.0f;
+                this.leftUpperArm.rotateAngleZ = -MathHelper.sin(onGround * 3.141593f) * -0.4f;
             }
             else
             {
                 float f6 = onGround;
-                this.body.rotateAngleY = MathHelper.sin(MathHelper.sqrt_float(f6) * 3.141593F * 2.0F) * 0.2F;
-                this.rightUpperArm.rotationPointZ = MathHelper.sin(this.body.rotateAngleY) * 2.5F;
-                this.leftUpperArm.rotationPointZ = -MathHelper.sin(this.body.rotateAngleY) * 2.5F;
+                this.body.rotateAngleY = MathHelper.sin(MathHelper.sqrt_float(f6) * 3.141593f * 2.0f) * 0.2f;
+                this.rightUpperArm.rotationPointZ = MathHelper.sin(this.body.rotateAngleY) * 2.5f;
+                this.leftUpperArm.rotationPointZ = -MathHelper.sin(this.body.rotateAngleY) * 2.5f;
                 this.rightUpperArm.rotateAngleY += this.body.rotateAngleY;
                 this.leftUpperArm.rotateAngleY += this.body.rotateAngleY;
                 this.leftUpperArm.rotateAngleX += this.body.rotateAngleY;
-                f6 = 1.0F - onGround;
+                f6 = 1.0f - onGround;
                 f6 *= f6;
                 f6 *= f6;
-                f6 = 1.0F - f6;
-                float f8 = MathHelper.sin(f6 * 3.141593F);
-                float f10 = MathHelper.sin(onGround * 3.141593F) * -(this.head.rotateAngleX - 0.7F) * 0.75F;
-                this.rightUpperArm.rotateAngleX -= (double)f8 * 1.2D + (double)f10;
-                this.rightUpperArm.rotateAngleY += this.body.rotateAngleY * 2.0F;
-                this.rightUpperArm.rotateAngleZ = MathHelper.sin(onGround * 3.141593F) * -0.4F;
+                f6 = 1.0f - f6;
+                float f8 = MathHelper.sin(f6 * 3.141593f);
+                float f10 = MathHelper.sin(onGround * 3.141593f) * -(this.head.rotateAngleX - 0.7f) * 0.75f;
+                this.rightUpperArm.rotateAngleX -= f8 * 1.2d + f10;
+                this.rightUpperArm.rotateAngleY += this.body.rotateAngleY * 2.0f;
+                this.rightUpperArm.rotateAngleZ = MathHelper.sin(onGround * 3.141593f) * -0.4f;
             }
         }
-        this.rightUpperArm.rotateAngleZ += MathHelper.cos(f2 * 0.09F) * 0.02F + 0.02F;
-        this.leftUpperArm.rotateAngleZ -= MathHelper.cos(f2 * 0.09F) * 0.02F + 0.02F;
+        this.rightUpperArm.rotateAngleZ += MathHelper.cos(f2 * 0.09f) * 0.02f + 0.02f;
+        this.leftUpperArm.rotateAngleZ -= MathHelper.cos(f2 * 0.09f) * 0.02f + 0.02f;
         
         if ( npc.getMode() != null && npc.isMode(EnumMode.SIT) )
         {
@@ -230,12 +230,12 @@ public class ModelSlim extends ModelBase implements IModelWithHands
 /* // 足をぶらぶらさせる感じ...余裕があったら設定で切り替え可能にしよう...
         if ( npc.isSitting() )
         {
-            this.rightUpperArm.rotateAngleX += -((float)Math.PI / 5F);
-            this.leftUpperArm.rotateAngleX += -((float)Math.PI / 5F);
-            this.rightUpperLeg.rotateAngleX = -((float)Math.PI * 2F / 5F);
-            this.leftUpperLeg.rotateAngleX = -((float)Math.PI * 2F / 5F);
-            this.rightUpperLeg.rotateAngleY = ((float)Math.PI / 10F);
-            this.leftUpperLeg.rotateAngleY = -((float)Math.PI / 10F);
+            this.rightUpperArm.rotateAngleX += -((float)Math.PI / 5f);
+            this.leftUpperArm.rotateAngleX += -((float)Math.PI / 5f);
+            this.rightUpperLeg.rotateAngleX = -((float)Math.PI * 2f / 5f);
+            this.leftUpperLeg.rotateAngleX = -((float)Math.PI * 2f / 5f);
+            this.rightUpperLeg.rotateAngleY = ((float)Math.PI / 10f);
+            this.leftUpperLeg.rotateAngleY = -((float)Math.PI / 10f);
             
             this.body.rotationPointY = this.bodyRotationPointY + this.sittingOffset;
             this.rightUpperArm.rotationPointY = this.armsRotationPointY + this.sittingOffset;
